Add files via upload
This is an API to data volumes from the SIEM platform Netwitness RSA. It organizes the data according the the component the data was collected from and the organization that the data belongs too which is enter using parsed arguments when running the script. It runs only once and waits for 2 AM every morning and run automatically to collect data volumes for the previous day for every hour for all organization organized by every component. This data is than sent to a file which is monitored by Splunk. And further analyses can happen to the data.